Conflicts:
	titan-cassandra/src/main/java/com/thinkaurelius/titan/diskstorage/cassandra/thrift/CassandraThriftKeyColumnValueStore.java
	titan-core/src/main/java/com/thinkaurelius/titan/diskstorage/keycolumnvalue/CachedKeyColumnValueStore.java
	titan-core/src/main/java/com/thinkaurelius/titan/diskstorage/locking/AbstractLocker.java
	titan-core/src/main/java/com/thinkaurelius/titan/diskstorage/util/MetricInstrumentedStore.java
	titan-core/src/main/java/com/thinkaurelius/titan/graphdb/database/EdgeSerializer.java
	titan-core/src/main/java/com/thinkaurelius/titan/graphdb/database/cache/ExpirationStoreCache.java
	titan-core/src/main/java/com/thinkaurelius/titan/graphdb/transaction/StandardTitanTx.java
	titan-test/src/test/java/com/thinkaurelius/titan/diskstorage/locking/ConsistentKeyLockerTest.java

+++ titan-dist/src/pkg/static/pkgcommon/bin/clean.sh
@@ -10,5 +10,7 @@ cd "`dirname $0`/../../"
 . pkgcommon/etc/config.sh
 
 # Clean dependency dir
-echo Deleting contents of "$DEP_DIR"...
-find $DEP_DIR -mindepth 1 -maxdepth 1 -exec rm -rf '{}' \;
+if [ -e "$DEP_DIR" ]; then
+    echo Deleting contents of "$DEP_DIR"...
+    find $DEP_DIR -mindepth 1 -maxdepth 1 -exec rm -rf '{}' \;
+fi
